Time: 2 ms (56.06%) | Memory: 8.1 MB (4.23%) - LeetSync

+++ 374-guess-number-higher-or-lower/guess-number-higher-or-lower.cpp
@@ -11,9 +11,9 @@ class Solution {
 public:
     int guessNumber(int n) {
 
-        int left=1;
-        int right=n;
-        int mid;
+        long long left=1;
+        long long right=n;
+        long long mid;
 
         while(left<=right){
             mid=left+((right-left)/2);
